Timelag
Added a timelag-transform.

This transform requires data from before the queried data, so I had to
pass the RHS path to the transform function so it can query for it. This
should be fine, I thought it would make writing tests a little harder,
but it wasn't a problem.

The tlag works great! Now we can start writing more transforms.