Rewrite from POC
This branch is a rewrite of the data transform service, now that we're
going from the POC stage, to the alpha production stage. The rewrite
aims to carry over the ideas that were proven to work in the POC, while
hardening them, tidying up the code, and writing tests.

We also aim to include all of the transforms from V2, as well as new
ones written for V3.

This first commit introduces a new project structure which makes more
sense. In addition, much dead code was deleted.

It is important to modularize functionality, making it easier to write
tests. Configuration (side-effect) should only have an effect in runtime
code, which is not tested, all other code should be completely covered.